+++ heylogs-cli/src/main/java/nbbrd/heylogs/cli/InitCommand.java
@@ -2,7 +2,6 @@ package nbbrd.heylogs.cli;
 
 import com.vladsch.flexmark.util.ast.Document;
 import internal.heylogs.cli.ConfigOptions;
-import internal.heylogs.cli.DebugOptions;
 import internal.heylogs.cli.SpecialProperties;
 import nbbrd.heylogs.Config;
 import nbbrd.heylogs.Heylogs;
@@ -46,9 +45,6 @@ public final class InitCommand implements Callable<Void> {
 
     @CommandLine.Mixin
     private ConfigOptions configOptions;
-
-    @CommandLine.Mixin
-    private DebugOptions debugOptions;
 
     @Override
     public Void call() throws Exception {