+++ heylogs-cli/src/main/java/nbbrd/heylogs/cli/ReleaseCommand.java
@@ -3,7 +3,6 @@ package nbbrd.heylogs.cli;
 import com.vladsch.flexmark.util.ast.Document;
 import internal.heylogs.cli.ChangelogInputParameters;
 import internal.heylogs.cli.ConfigOptions;
-import internal.heylogs.cli.DebugOptions;
 import internal.heylogs.cli.SpecialProperties;
 import nbbrd.heylogs.Config;
 import nbbrd.heylogs.Heylogs;
@@ -43,9 +42,6 @@ public final class ReleaseCommand implements Callable<Void> {
 
     @CommandLine.Mixin
     private ConfigOptions configOptions;
-
-    @CommandLine.Mixin
-    private DebugOptions debugOptions;
 
     @Override
     public Void call() throws Exception {